// angular/src/openapi/model/registerRequest.ts
/**
 * ProLinked Web Api
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface RegisterRequest { 
    name: string;
    surname: string;
    userName: string;
    email: string;
    password: string;
    dateOfBirth: Date;
    phoneNumber?: string;
    summary?: string;
    jobTitle?: string;
    company?: string;
    city?: string;
}